refactor cache and files. Tests for ACM & IAM still missing.
